# proxy/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Cost Control Tower — Ingestion API
  version: "1.0.0"
  license:
    name: Internal use only — not for external distribution
  description: |
    ## Getting started

    This API lets an internal HCLTech application write its own AI usage and
    outcome events into the same Cost Control Tower tables Product Studio
    already uses, so it can be traced, costed, and governed the same way.

    **Registration is not self-service.** Contact the platform operator to
    register your app and receive a credential — registration and credential
    issuance are direct SQL/RPC, run manually, not exposed over HTTP
    (see the spec's Section 3/4 for why).

    **Credential handling.** Your API key is shown exactly once, at issuance
    time. Move it into your own team's secrets manager immediately — it is
    never stored in plaintext or recoverable after that point. Every request
    authenticates with `Authorization: Bearer <api_key>`; `company_id` and
    `app_id` are resolved from the credential server-side and are never
    accepted as request fields.

    ```bash
    curl -X POST https://<ingestion-host>/v1/usage-events \
      -H "Authorization: Bearer <your_api_key>" \
      -H "Content-Type: application/json" \
      -d '{
        "client_call_id": "a1b2c3d4-...",
        "user_role_at_call": "engineer",
        "caller": "my-app-feature-x",
        "requested_model": "gpt-4o",
        "status": "success",
...
      }'
    ```

...

...
      shape.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /v1/usage-events:
    post:
      operationId: createUsageEvents
      tags: [usage-events]
      summary: Record one or a batch of AI usage events
      description: |
        Accepts a single object or an array (batch cap 500 items). Each item
        is validated and inserted independently — a single malformed item
        never discards the rest of the batch. `company_id`/`app_id` are
        resolved from the credential.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UsageEventInput'
                - type: array
                  maxItems: 500
                  items:
                    $ref: '#/components/schemas/UsageEventInput'
      responses:
        '200':
          description: |
            Single-item request: a plain result object. Batch request: a
            `results` array, one entry per submitted item, in submitted
            order — `index` maps back to the item's position in the request
            array. The request itself still returns 400 if it's structurally
            invalid, independent of any individual item's validity.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/IngestResult'
                  - $ref: '#/components/schemas/BatchIngestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      operationId: listUsageEvents
      tags: [usage-events]
      summary: Read back your own submitted usage events
      description: |
        Reconciliation, not a dashboard replacement. Returns only the
        caller's own events (scoped by the credential-resolved
        `company_id`/`app_id`), including a computed `calculated_cost` per
        event. Ordered `request_started_at DESC, id DESC`.
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 200, minimum: 1, maximum: 1000 }
        - name: cursor
          in: query
          required: false
          description: |
            Opaque to the caller — a base64-encoded `(request_started_at,
            id)` tuple of the last row in the prior page. Pass back the
            `next_cursor` value from a previous response verbatim.
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsageEventRecord'
                  next_cursor:
                    type: string
                    nullable: true
                    description: Pass as `cursor` to fetch the next page. Null when this is the last page.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/usage-events/{client_call_id}/units-generated:
    patch:
      operationId: reportUnitsGenerated
      tags: [usage-events]
      summary: Report a variable unit count after parsing your own response
      description: |
        Idempotent, scoped to `(company_id, app_id, client_call_id)`. Only
        meaningful for `yield_ratio`-costed outcome types — constant-per-call
        counts should go on the initial write instead.
      parameters:
        - name: client_call_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [units_generated]
              properties:
                units_generated: { type: integer, minimum: 0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  units_generated: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The API key issued to your app, as a Bearer token.
  responses:
    BadRequest:
      description: Malformed payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Missing, invalid, or inactive credential
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    NotFound:
      description: Referenced resource doesn't exist (or doesn't belong to the caller)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            type: { type: string }
            message: { type: string }
    IngestResult:
      type: object
      properties:
        id: { type: string }
        deduplicated: { type: boolean }
    BatchIngestResponse:
      type: object
      properties:
        results:
          type: array
          items:
            oneOf:
              - type: object
                required: [index, id, deduplicated]
                properties:
                  index: { type: integer }
                  id: { type: string }
                  deduplicated: { type: boolean }
              - type: object
                required: [index, error]
                properties:
                  index: { type: integer }
                  error: { $ref: '#/components/schemas/ErrorBody' }
    UsageEventInput:
      type: object
      required: [client_call_id, user_role_at_call, caller, requested_model, status, request_started_at]
      properties:
        client_call_id: { type: string, format: uuid, description: "Your own UUID, idempotency key" }
        provider: { type: string, default: anthropic }
        product_id: { type: string, format: uuid, nullable: true, description: "Omit for non-Product-Studio apps" }
        session_id: { type: string, nullable: true, description: "Your own stable identifier grouping related actions" }
        user_id: { type: string, nullable: true, description: "Plain UUID, no FK, use your own internal identifier" }
        user_role_at_call: { type: string, description: "Free text, open" }
        caller: { type: string, description: "Your own internal name for what triggered this call" }
        requested_model: { type: string }
        response_model: { type: string, nullable: true }
        settings_mode:
          type: string
          nullable: true
          description: "Defaults to 'external' if omitted. Only set explicitly if a Product-Studio-internal value genuinely applies."
        selection_rule:
          type: string
          nullable: true
          description: "Same default behavior as settings_mode."
        input_tokens: { type: integer, nullable: true }
        output_tokens: { type: integer, nullable: true }
        cache_creation_5m_tokens: { type: integer, nullable: true }
        cache_creation_1h_tokens: { type: integer, nullable: true }
        cache_read_tokens: { type: integer, nullable: true }
        provider_usage_raw: { type: object, nullable: true, description: "jsonb escape hatch" }
        status: { type: string, enum: [success, error, timeout] }
        provider_http_status: { type: integer, nullable: true }
        error_type: { type: string, nullable: true }
        failure_phase: { type: string, nullable: true, description: "Product-Studio-specific enum, leave null if it doesn't apply" }
        request_started_at: { type: string, format: date-time }
        duration_ms: { type: integer, nullable: true }
        request_bytes: { type: integer, nullable: true }
        response_bytes: { type: integer, nullable: true }
        outcome_id: { type: string, format: uuid, nullable: true, description: "Set if this event belongs to an outcome created via /v1/outcomes" }
        units_generated: { type: integer, nullable: true, description: "Only meaningful for yield_ratio-costed outcome types" }
        client_trace_id:
          type: string
          nullable: true
          description: "Your own trace-continuation key (AI Trace Layer). Generate it once per conversation/agent run and send it on every call belonging to that run, forever — never a server-generated id. Creates the trace implicitly on first use if you haven't called POST /v1/traces first. Requires agent_name."
        agent_name:
          type: string
          nullable: true
          description: "Required if client_trace_id is present. Your own name for the agent/flow this call belongs to."
    UsageEventRecord:
      allOf:
        - $ref: '#/components/schemas/UsageEventInput'
        - type: object
          properties:
            id: { type: string }
            trace_id: { type: string, format: uuid, nullable: true }
            calculated_cost:
              type: number
              nullable: true
              description: "Null when no matching mt_model_pricing row was found for this provider/model/date."
...